apply validation

+++ src/controllers/user/user.controller.ts
@@ -6,8 +6,9 @@ import {
 	GraphQLString,
 } from "graphql";
 import PaginationType from "../../type-defs/pagination.type";
-import UserType from "../../type-defs/user.type";
+import {UserType} from "../../type-defs/user.type";
 import UserService from "./user.service";
+import validator from "validator"
 
 export const userPagination = {
 	type: PaginationType,
@@ -66,11 +67,16 @@ export const getUser = {
 export const storeUser = {
 	type: UserType,
 	args: {
-		name: { type: GraphQLString },
+		name: {type: GraphQLString}
 	},
-	resolve: async (_: any, { name }: any) => {
+	resolve: async (_: any, { name }: any) => {		
 		try {
+			if (validator.isEmpty(name)) {
+				throw new Error("The name field must not be empty.")
+			}
+
 			const result = await new UserService().store({ name });
+			
 			return result;
 		} catch (error) {
 			throw new Error(error);
@@ -82,12 +88,28 @@ export const updateUser = {
 	type: UserType,
 	args: {
 		id: { type: GraphQLID! },
-		name: { type: GraphQLNonNull(GraphQLString)! },
+		name: { type: GraphQLString! },
 	},
 	resolve: async (_: any, { id, name }: any) => {
 		try {
-			const result = await new UserService().update({ name }, id);
-			return await new UserService().findOne(id);
+			if (validator.isEmpty(id)) {
+				throw new Error("The id field must not be empty.");
+			}
+
+			const isExist = await new UserService().findOne(id);
+			if (!isExist) {
+				throw new Error("The user with specified ID is doesn't exist.");
+			}
+
+			if (validator.isEmpty(name)) {
+				throw new Error("The name field must not be empty.");
+			}
+
+			await new UserService().update({ name }, id);
+
+			const user = await new UserService().findOne(id);
+			
+			return user;
 		} catch (error) {
 			throw new Error(error);
 		}
@@ -101,7 +123,16 @@ export const deleteUser = {
 	},
 	resolve: async (_: any, { id }: any) => {
 		try {
+			if (validator.isEmpty(id)) {
+				throw new Error("The user with specified ID is doesn't exist.");
+			}
+
 			const result = await new UserService().delete(id);
+			
+			if (!result.affected) {
+				throw new Error("The user with specified ID is doesn't exist.");
+			}
+
 			return true;
 		} catch (error) {
 			throw new Error(error);
